Android: Made popup and menu windows work, fixed Fl::wait() recursion,          fixed incompatible Fl::wait() behavior.
git-svn-id: file:///fltk/svn/fltk/branches/branch-1.4@12793 ea41ed52-d2ee-0310-a9c1-e6b18d33e121

+++ src/drivers/Android/Fl_Android_Graphics_Driver.cxx
@@ -892,10 +892,10 @@ void Fl_Android_Graphics_Driver::pie_unscaled(float xi, float yi, float w, float
  * FIXME: these do not draw rotated ellipses correctly!
  * */
 void Fl_Android_Graphics_Driver::ellipse_unscaled(double xt, double yt, double rx, double ry) {
-  double llx = xt-rx;
-  double w = xt+rx-llx;
-  double lly = yt-ry;
-  double h = yt+ry-lly;
+  float llx = xt-rx;
+  float w = xt+rx-llx;
+  float lly = yt-ry;
+  float h = yt+ry-lly;
 
   if (what==POLYGON)
     pie_unscaled(llx, lly, w, h, 0.0, 360.0);
